[tests] Cleanup Suite: upd
