+++ integration/cleanup/_fixtures/cleanup_unused_stages/werf.yaml
